Fixes exception in "compatibility" PSAT,GBS.
With --format-level=0, the lat/lon/alt error is reported even when all
zero, but this was also attempting to output it when all are None,
causing an exception.  This was only noticeable in --stop-on-error
mode.  The fix is to ignore the exception.

TESTED:
The tests now pass, even with -s.