Reworks GBS handling.
The GBS code was originally written for the Hemisphere proprietary
PSAT,GBS sentence.  There is now a standard GxGBS sentence, but it
differs slightly in that it lacks Hemisphere's 'flag' field.  This
affects the parser and formatter, but not the decoder (since the
decoder wasn't processing the flag field, anyway).  The reworked
code has both generic GxGBS support and PSAT,GBS support.

TESTED:
The usual tests pass.  No data with GxGBS yet.