Reworks extracter length limit.
This switches from the original scale-factor-based approach to a fixed
limit (overridable on a per-type basis) for more consistency across
formats.  In the Hemisphere case, it includes a check to ensure that
the limit isn't smaller than the largest known length.  This may or
may not be applicable to other formats.

The length check is also moved before the "needed" check, since the
badness of the length shouldn't depend on whether a large amount of
data has already been fetched.

TESTED:
The usual tests pass.