Adds dummy Garmin support.
The Garmin extracter needs to be somewhat more complicated due to the
DLE doubling stuff, and there's currently no available test data.  So
for now, this is just a nonfunctional skeleton.

This is only being added now because it's convenient to do so when
there's a minimal implementation for another protocol available for
copying.  Since it's not currently useful at all, it's not yet added
to the 'combined' modules.