Fixes bug in Hemisphere binary extracter.
The code rejecting overly long data was a holdover from an earlier
version of the buffer management where this wasn't expected.  Now,
it's entirely possible to have extra data in the buffer due to
other extracters pulling it in.

TESTED:
Usual regression tests pass, with the new Garmin support that
triggered this bug.