Updates NMEA signal decoding for multi-GNSS.
This switches to the new two-level signal decoding, which takes the
system into account.  Because it changes the old (incorrect) GLONASS
behavior, it includes a format-level-based fallback to the old
behavior.

TESTED:
The usual tests pass, based on their respective format levels.
The four-GNSS sample from the u-blox 9 decodes correctly.